delete some examples, migrate logo and loading spinner as proper components and integrate them into component factory, cleaning up
