+++ src/main/webapp/js/GEPPETTO.js
@@ -586,8 +586,6 @@ define(function (require) {
     require('model/ProjectFactory')(GEPPETTO);
     require('model/ModelFactory')(GEPPETTO);
     require('model/ExperimentsController')(GEPPETTO);
-    
-    require('components/ComponentFactory')(GEPPETTO);
 
     return GEPPETTO;
