+++ src/main/webapp/js/components/ComponentsInitialization.js
@@ -33,150 +33,14 @@
 
 
 define(function (require) {
-	
 	return function (GEPPETTO) {
-
-		GEPPETTO.showExecutionDialog = function(callback){
-			var formCallback = callback;
-			
-			var formId = "gptForm";
-			
-			var formName = "Simulation Form";
-			
-			var schema = {
-				  type: "object",
-				  required: ["experimentName", "timeStep", "length", "simulator", "numberProcessors"],
-				  properties: {
-					experimentName: {type: "string", title: "Experiment Name"},
-					timeStep:{type:'number', title: 'Time Step (s)'},
-					length:{type:'number', title: 'Length (s)'},
-					simulator:{
-					      type: "string",
-					      title: "Simulator",
-					      enum: ["neuronSimulator", "lemsSimulator", "neuronNSGSimulator"],
-					      enumNames: ["Neuron", "jLems", "Neuron on NSG"]
-					},
-	
-					numberProcessors:{type:'number', title: 'Number of Processors'}
-				  }
-				};
-	
-			var formData= {
-					experimentName: Project.getActiveExperiment().getName(),
-					timeStep: Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getTimeStep(),
-					length: Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getLength(),
-					simulator:Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getSimulator(),
-					numberProcessors: 1
-			};
-			
-			var submitHandler = function(){
-				GEPPETTO.Flows.showSpotlightForRun(formCallback);
-				$("#" + formWidget.props.id + "_container").remove();
-			};
-			
-			var errorHandler = function(){
-				
-			};
-			
-			var changeHandler = function(formObject){
-				for (var key in formObject.formData) {
-					if (formObject.formData[key] != this.formData[key]){
-						if (key == 'experimentName'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='name']").html(formObject.formData[key]).blur();
-						}
-						else if (key == 'timeStep'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='timeStep']").html(formObject.formData[key]).blur();
-						}
-						else if (key == 'length'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='length']").html(formObject.formData[key]).blur();					
-						}
-						else if (key == 'simulator'){
-							Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].setSimulatorParameter('numberProcessors', formObject.formData[key]).blur();
-						}
-						else if (key == 'numberProcessors'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='simulatorId']").html(formObject.formData[key]).blur();
-						}
-						this.formData[key] = formObject.formData[key];
-					}
-				}
-			};
-	
-			var formWidget = GEPPETTO.ComponentFactory.addComponent('FORM',
-					{id: formId, name:formName, schema:schema, formData:formData, submitHandler:submitHandler, errorHandler:errorHandler, changeHandler:changeHandler});
-		};
-	
-		GEPPETTO.showingPanelConcept = function(){
-			var formId = "gptForm";
-			
-			var formName = "Simulation Form";
-			
-			var schema = {
-				  type: "object",
-				  required: ["experimentName", "timeStep", "length", "simulator", "numberProcessors"],
-				  properties: {
-					experimentName: {type: "string", title: "Experiment Name"},
-					timeStep:{type:'number', title: 'Time Step (s)'},
-					length:{type:'number', title: 'Length (s)'},
-					simulator:{
-					      type: "string",
-					      title: "Simulator",
-					      enum: ["neuronSimulator", "lemsSimulator", "neuronNSGSimulator"],
-					      enumNames: ["Neuron", "jLems", "Neuron on NSG"]
-					},
-	
-					numberProcessors:{type:'number', title: 'Number of Processors'}
-				  }
-				};
-	
-			var formData= {
-					experimentName: Project.getActiveExperiment().getName(),
-					timeStep: Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getTimeStep(),
-					length: Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getLength(),
-					simulator:Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getSimulator(),
-					numberProcessors: 1
-			};
-			
-			var submitHandler = function(){
-				GEPPETTO.Flows.showSpotlightForRun(formCallback);
-				$("#" + formWidget.props.id + "_container").remove();
-			};
-			
-			var errorHandler = function(){
-				
-			};
-			
-			var changeHandler = function(formObject){
-				for (var key in formObject.formData) {
-					if (formObject.formData[key] != this.formData[key]){
-						if (key == 'experimentName'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='name']").html(formObject.formData[key]).blur();
-						}
-						else if (key == 'timeStep'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='timeStep']").html(formObject.formData[key]).blur();
-						}
-						else if (key == 'length'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='length']").html(formObject.formData[key]).blur();					
-						}
-						else if (key == 'simulator'){
-							Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].setSimulatorParameter('numberProcessors', formObject.formData[key]).blur();
-						}
-						else if (key == 'numberProcessors'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='simulatorId']").html(formObject.formData[key]).blur();
-						}
-						this.formData[key] = formObject.formData[key];
-					}
-				}
-			};
-			
-			var panelChildren = [];
-			panelChildren.push(GEPPETTO.ComponentFactory.getComponent('DIVFEO',{id:"pp", text:'xaka'}));
-			panelChildren.push(GEPPETTO.ComponentFactory.getComponent('FORM',{id: "pp2", name:"mm", schema:schema, formData:formData, submitHandler:submitHandler, errorHandler:errorHandler, changeHandler:changeHandler}));
-			
-			var panelComponent = GEPPETTO.ComponentFactory.addComponent('PANEL',	{id: "kklr", name:"pl"});
-			panelComponent.addChildren(panelChildren);
-		};
-	
+		//Logo initialization 
+		GEPPETTO.ComponentFactory.addComponent('LOGO', {logo: 'gpt-gpt_logo'}, document.getElementById("geppettologo"));
 		
-		GEPPETTO.Flows.addCompulsoryAction('GEPPETTO.showExecutionDialog', GEPPETTO.Resources.PLAY_FLOW);
+		//Loading spinner initialization
+		GEPPETTO.on('show_spinner', function(label) {
+			GEPPETTO.ComponentFactory.addComponent('LOADINGSPINNER', {show : true, keyboard : false, text: label, logo: "gpt-gpt_logo"}, document.getElementById("modal-region"));	
+		});
+		
 	};
 });
