+++ src/main/webapp/js/components/OSB.css
@@ -1,0 +1,6 @@
+.spinner-container .gpt-osb {
+  width: 100px;
+  font-size: 100px;
+  margin: 0 auto;
+  margin-bottom: 30px;
+}
