+++ src/main/webapp/js/components/components.js
@@ -6,7 +6,7 @@ require('jsx!./dev/home/HomeControl');
 require('jsx!./dev/simulationcontrols/ExperimentControls');
 require('jsx!./dev/cameracontrols/CameraControls');
 //require('jsx!./dev/tutorial/IntroModal');
-require('jsx!./dev/logo/logo');
+//require('jsx!./dev/logo/logo');
 require('jsx!./dev/share/share');
 //require('jsx!./dev/contact/contact');
 require('jsx!./dev/foregroundcontrols/ForegroundControls');
